move AgentSettings out of AgentSettingsResolver
